Replaced all fetch functions to APIClient.js, fixed all routes and imports
